@@ -1,24 +1,18 @@
 import { TestBed } from '@angular/core/testing';
 import { AppComponent } from './app.component';
-import { NxWelcomeComponent } from './nx-welcome.component';
 import { RouterTestingModule } from '@angular/router/testing';
+import { ReactiveFormsModule } from '@angular/forms';
+
 describe('AppComponent', () => {
 	beforeEach(async () => {
 		await TestBed.configureTestingModule({
-			imports: [AppComponent, NxWelcomeComponent, RouterTestingModule],
+			imports: [AppComponent, RouterTestingModule, ReactiveFormsModule],
 		}).compileComponents();
 	});
-	it('should render title', () => {
-		const fixture = TestBed.createComponent(AppComponent);
-		fixture.detectChanges();
-		const compiled = fixture.nativeElement as HTMLElement;
-		expect(compiled.querySelector('h1')?.textContent).toContain(
-			'Welcome angular-loaders'
-		);
-	});
-	it(`should have as title 'angular-loaders'`, () => {
+
+	it('should create the app', () => {
 		const fixture = TestBed.createComponent(AppComponent);
 		const app = fixture.componentInstance;
-		expect(app.title).toEqual('angular-loaders');
+		expect(app).toBeTruthy();
 	});
 });
